requirements.txt
mysql.connector>=8.0.30
pandas>=1.3.4
numpy>=1.20.3
matplotlib>=3.4.3
seaborn>=0.11.2
streamlit==1.10.0
sklearn>=0.24.2
pickle>=4.0